Book Proposal
This is the publisher's standard book proposal form. Normal text is their questions, italics are my answers. Comments within [brackets] are only for Tilt internally, not to be sent to the publisher.

I'd like feedback on whether we as a company are happy with this, and if not, where to improve. Please let me know if anything requires further clarification, and I'm open to having a call to discuss things in detail if necessary.

Please don't mention the publisher's name in public for now :)